+++ perl/FSM/Pipeline/SourceGenerationOrchestrator.pm
@@ -57,6 +57,7 @@ sub generate_from_file ($class, %args) {
         my $source_info = FSM::Pipeline::SourceFrontend->classify_source_ast($raw_ast);
 
         FSM::Pipeline::SourceFrontend->enforce_strict_source_boundary(
+            raw_ast => $raw_ast,
             strict_mode => ($pipeline->{strict_mode} // 0),
             source_info => $source_info,
             source_label => $fsm_file,